requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
ciso8601==2.2.0
colorlog==6.6.0
commonmark==0.9.1
Django==4.0.6
frozenlist==1.3.0
hikari==2.0.0.dev109
hikari-lightbulb==2.2.4
idna==3.3
image==1.5.33
multidict==6.0.2
Pillow==9.2.0
pycares==4.2.1
pycparser==2.21
Pygments==2.12.0
pyimgur==0.6.0
pytz==2022.1
qrcode==7.3.1
requests==2.28.1
rich==12.5.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
urllib3==1.26.11
yarl==1.7.2
